+++ flatlaf-core/src/main/java/com/formdev/flatlaf/ui/FlatPopupFactory.java
@@ -32,7 +32,6 @@ import java.awt.Panel;
 import java.awt.Point;
 import java.awt.PointerInfo;
 import java.awt.Rectangle;
-import java.awt.Toolkit;
 import java.awt.Window;
 import java.awt.event.ComponentEvent;
 import java.awt.event.ComponentListener;
@@ -312,7 +311,7 @@ public class FlatPopupFactory
 			return null;
 
 		Rectangle screenBounds = gc.getBounds();
-		Insets screenInsets = Toolkit.getDefaultToolkit().getScreenInsets( gc );
+		Insets screenInsets = FlatUIUtils.getScreenInsets( gc );
 		int screenTop = screenBounds.y + screenInsets.top;
 
 		// place tooltip above mouse location if there is enough space
@@ -547,7 +546,15 @@ public class FlatPopupFactory
 			int x = popupWindow.getX();
 			int y = popupWindow.getY();
 
-			popup.show();
+			if( !popupWindow.isVisible() )
+				popup.show();
+			else {
+				// if the popup window is already visible (because it is reused),
+				// do not invoke Popup.show() because this would invoke Window.toFront(),
+				// which may have the side effect that an inactive owner window
+				// would be also moved to front and maybe hide previously active window
+				popupWindow.pack();
+			}
 
 			// restore popup window location if it has changed
 			// (probably scaled when screens use different scale factors)
@@ -619,7 +626,30 @@ public class FlatPopupFactory
 
 		void showImpl() {
 			if( delegate != null ) {
-				showPopupAndFixLocation( delegate, popupWindow );
+				// On macOS and Linux, the empty popup window is shown in popup.show()
+				// (in peer.setVisible(true) invoked from Component.show()),
+				// but the popup content is painted later via repaint manager.
+				// This may cause some flicker, especially during JVM warm-up or
+				// when running JVM in interpreter mode (option -Xint).
+				// To reduce flicker, immediately paint popup content as soon as popup window becomes visible.
+				// This also fixes a problem with JetBrainsRuntime JVM, where sometimes the popups were empty.
+				if( (SystemInfo.isMacOS || SystemInfo.isLinux) && popupWindow instanceof JWindow ) {
+					HierarchyListener l = e -> {
+						if( e.getID() == HierarchyEvent.HIERARCHY_CHANGED &&
+							(e.getChangeFlags() & HierarchyEvent.SHOWING_CHANGED) != 0 )
+						{
+							((JWindow)popupWindow).getRootPane().paintImmediately(
+								0, 0, popupWindow.getWidth(), popupWindow.getHeight() );
+						}
+					};
+					popupWindow.addHierarchyListener( l );
+					try {
+						showPopupAndFixLocation( delegate, popupWindow );
+					} finally {
+						popupWindow.removeHierarchyListener( l );
+					}
+				} else
+					showPopupAndFixLocation( delegate, popupWindow );
 
 				// increase tooltip size if necessary because it may be too small on HiDPI screens
 				//    https://bugs.openjdk.java.net/browse/JDK-8213535
@@ -678,6 +708,21 @@ public class FlatPopupFactory
 				// restore background so that it can not affect other LaFs (when switching)
 				// because popup windows are cached and reused
 				popupWindow.setBackground( oldPopupWindowBackground );
+
+				// On macOS, popupWindow.setBackground(...) invoked from constructor,
+				// has no affect if the popup window peer (a NSWindow) is already created,
+				// which is the case when reusing a cached popup window
+				// (see class PopupFactory.HeavyWeightPopup).
+				// This may result in flicker when e.g. showing a popup in a light theme,
+				// then switching to a dark theme and again showing a popup,
+				// because the underling NSWindow still has a light background,
+				// which may be shown shortly before the actual dark popup content is shown.
+				// To fix this, dispose the popup window, which disposes the NSWindow.
+				// The AWT popup window stays in the popup cache, but when reusing it later,
+				// a new peer and a new NSWindow is created and gets the correct background.
+				if( SystemInfo.isMacOS )
+					popupWindow.dispose();
+
 				popupWindow = null;
 			}
 		}